server initialized, changed directories
