+++ client/src/main/java/pl/pwsztar/Connect/Customer.java
