+++ client/src/main/java/pl/pwsztar/Connect/CustomerDto.java
