+++ client/src/main/resources/pl/pwsztar/deactivate.fxml
