+++ client/src/main/resources/pl/pwsztar/payLoan.fxml
