+++ client/src/main/resources/pl/pwsztar/payOut.fxml
