+++ client/src/main/resources/pl/pwsztar/registerVerification.fxml
