+++ client/src/main/resources/pl/pwsztar/takeLoan.fxml
