+++ client/src/main/resources/pl/pwsztar/transfer.fxml
