+++ server/src/main/java/pl/pwsztar/rest/ApiController.java
@@ -1,0 +1,18 @@
+package pl.pwsztar.rest;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/bank")
+public class ApiController {
+
+    @GetMapping(value = "test")
+    public ResponseEntity<Void> testServer() {
+        return new ResponseEntity<>(HttpStatus.OK);
+    }
+
+}
