+++ server/src/test/java/pl/pwsztar/ApplicationTests.java
@@ -1,0 +1,13 @@
+package pl.pwsztar;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
